Rename API settings module to envs
